requirements.txt
flask==3.1.0
gunicorn==23.0.0
garminconnect==0.2.40
requests-oauthlib==1.3.1
gspread==6.1.2
google-auth>=2.22.0
